Fixed several issues

+++ apps/web/src/api-wrapper/users.ts
@@ -5,8 +5,8 @@ import {
   type User
 } from "types-custom"
 import { fetchHandler } from "./fetchHandler"
-import Query from "./query"
 import { type APIResponse } from "./types"
+import { validate as isValidUUID } from "uuid"
 
 const USERS_ENDPOINT = "users"
 
@@ -21,11 +21,7 @@ export async function getUser(id: string): Promise<APIResponse<User>> {
 export async function getAllUsersPaged(
   page?: number
 ): Promise<APIResponse<GetAllPaginatedUserDto>> {
-  const query = new Query({
-    page
-  })
-
-  return await fetchHandler(`${USERS_ENDPOINT}${query.toString()}`)
+  return await fetchHandler(`${USERS_ENDPOINT}`, undefined, { page })
 }
 
 export async function createUser(
@@ -41,6 +37,13 @@ export async function updateUser(
   id: string,
   updateUserDto: UpdateUserDto
 ): Promise<APIResponse<User>> {
+  if (!isValidUUID(id)) {
+    return {
+      ok: false,
+      message: "Invalid UUID"
+    }
+  }
+
   return await fetchHandler(`${USERS_ENDPOINT}/${id}`, {
     method: "PATCH",
     body: JSON.stringify(updateUserDto)
@@ -48,6 +51,13 @@ export async function updateUser(
 }
 
 export async function deleteUser(id: string): Promise<APIResponse<User>> {
+  if (!isValidUUID(id)) {
+    return {
+      ok: false,
+      message: "Invalid UUID"
+    }
+  }
+
   return await fetchHandler(`${USERS_ENDPOINT}/${id}`, {
     method: "DELETE"
   })
